Avoid circular dependency, extract constants to a separate file.
