+++ src/coordinate-grid-mixin.js
@@ -3,7 +3,8 @@ import * as d3 from 'd3';
 import {colorMixin} from './color-mixin';
 import {marginMixin} from './margin-mixin';
 import {baseMixin} from './base-mixin';
-import {constants, optionalTransition, override, transition, units} from './core';
+import {optionalTransition, override, transition, units} from './core';
+import {constants} from './constants';
 import {utils} from './utils';
 import {logger} from './logger';
 import {filters} from './filters';
