+++ src/core.js
@@ -1,4 +1,5 @@
 import { utils } from './utils';
+import {constants} from './constants';
 
 /**
  * The entire dc.js library is scoped under the **dc** name space. It does not introduce
@@ -18,18 +19,6 @@ import { utils } from './utils';
  *      .filter('sunday');
  */
 export const version = '<%= conf.pkg.version %>';
-export const constants = {
-        CHART_CLASS: 'dc-chart',
-        DEBUG_GROUP_CLASS: 'debug',
-        STACK_CLASS: 'stack',
-        DESELECTED_CLASS: 'deselected',
-        SELECTED_CLASS: 'selected',
-        NODE_INDEX_NAME: '__index__',
-        GROUP_INDEX_NAME: '__group_index__',
-        DEFAULT_CHART_GROUP: '__default_chart_group__',
-        EVENT_DELAY: 40,
-        NEGLIGIBLE_NUMBER: 1e-10
-    };
 
 let _renderlet = null;
