+++ src/index.js
@@ -1,3 +1,4 @@
+export * from './constants';
 export * from './core';
 export * from './errors';
 export * from './utils';
